\name{DAAGxtras-package}
\alias{DAAGxtras-package}
\alias{DAAGxtras}
\docType{package}
\title{
The R DAAGxtras Package
}
\description{
various data sets used in additional exercises for the book
        Maindonald, J.H. and Braun, W.J. (3rd edn 2010)  
        "Data Analysis and Graphics Using R", and/or used for
        various instructional purposes.  
}
\details{
Note especially the datasets \code{rockArt} (multivariate binary data on
Pacific rock art), \code{audists} (Australian road distances),
\code{greatLakes} (Great Lakes lake levels, by year).
and the dataset \code{fishRivers} that has information on preferred
fish river types.

For a complete list, use \code{library(help="DAAGxtras")}.
}
\author{
Author: John Maindonald
  
Maintainer: John Maindonald  \email{john.maindonald@anu.edu.au}
}
\keyword{package}
