\name{DAAGxtras-package}
\alias{DAAGxtras-package}
\alias{DAAGxtras}
\docType{package}
\title{
The R DAAGxtras Package
}
\description{
various data sets used in additional exercises for the book
        Maindonald, J.H. and Braun, W.J. (2nd edn 2007)  
        "Data Analysis and Graphics Using R", and/or used for
        various instructional purposes.  
}
\details{
Note especially the datasets \code{nassCDS} (airbag and other
influences on US vehicle accident outcomes: 1997-2002),
\code{cricketer} (handedness, British first class cricketers
born 1840-1960 whose handedness could be determined),
\code{nswdemo} and related datasets (US Labor Training
Evaluation Data), \code{rockArt} (multivariate binary data on
Pacific rock art), \code{hotspots} (ages and distance from
Kilauea of larval hotspots; c.f. also \code{hotspots2006}),
\code{nihills} (record times for Northern Ireland mountain
races), \code{worldRecords} (record times for track and road
races), \code{progression} (progression of record times for
track races), \code{edcT} and \code{edcCO2} (temperature and CO2
estimates from the EPICA Dome C ice core, covering 0 to 800 kyr
BP), \code{grog} (Australian and NZ liquor consumption patterns,
1998-2006), and \code{bomregions} (historical Australian climate
data).

For a complete list, use \code{library(help="DAAGxtras")}.
}
\author{
Author: John Maindonald
  
Maintainer: John Maindonald  \email{john.maindonald@anu.edu.au}
}
\keyword{package}
